代码整理,与富友交互代码优化,独立交互代码

+++ src/main/java/com/gqhmt/funds/architect/trade/entity/FundTradeEntity.java
@@ -89,7 +89,7 @@ public class FundTradeEntity {
 	private Integer repaymentId;
 
 	@Column(name = "bonus_amount")
-	private Long bonusAmount;
+	private BigDecimal bonusAmount;
 
 	public FundTradeEntity() {
 	}
@@ -204,11 +204,11 @@ public class FundTradeEntity {
 	}
 
 	
-	public Long getBonusAmount() {
+	public BigDecimal getBonusAmount() {
 		return bonusAmount;
 	}
 
-	public void setBonusAmount(Long bonusAmount) {
+	public void setBonusAmount(BigDecimal bonusAmount) {
 		this.bonusAmount = bonusAmount;
 	}
 }
